added calculation to see user insights in admin mode, updated sync manager to avoid duplicating values when creating, removed user calculated properties like cursos completados and puntajes

+++ Coffey/Utils/Sync/SyncManager.swift
@@ -57,6 +57,9 @@ final class SyncManager {
         for local in updatedLocalArr {
             if let id = local.remoteID as? Int, id == 0 { continue }
 
+            // Skip records that were just created in Step 1
+            if synced.contains(local.remoteID) { continue }
+
             if local.deletedAt != nil {
                 if remoteByID[local.remoteID] != nil {
                     try await api.delete(local)
